# Add new shorthand data