--- a/lib/reference/shorthandData.js
+++ b/lib/reference/shorthandData.js
@@ -203,4 +203,14 @@ module.exports = {
     "text-emphasis-style",
     "text-emphasis-color",
   ],
+  "mask": [
+    "mask-image",
+    "mask-mode",
+    "mask-position",
+    "mask-size",
+    "mask-repeat",
+    "mask-origin",
+    "mask-clip",
+    "mask-composite",
+  ],
 }