--- a/lib/rules/declaration-block-no-redundant-longhand-properties/__tests__/index.js
+++ b/lib/rules/declaration-block-no-redundant-longhand-properties/__tests__/index.js
@@ -39,9 +39,6 @@ testRule(rule, {
     code: "a { border-top-width: 7px; border-top-style: double; border-top-color: green; }",
     message: messages.expected("border-top"),
   }, {
-    code: "a { -webkit-transition-delay: 0.5s; -webkit-transition-duration: 2s; -webkit-transition-property: top; -webkit-transition-timing-function: ease; }",
-    message: messages.expected("-webkit-transition"),
-  }, {
     code: "a { border-top-width: 1px; border-bottom-width: 1px; border-left-width: 1px; border-right-width: 1px; }",
     message: messages.expected("border-width"),
   } ],