--- a/lib/rules/declaration-block-no-redundant-longhand-properties/__tests__/index.js
+++ b/lib/rules/declaration-block-no-redundant-longhand-properties/__tests__/index.js
@@ -18,6 +18,12 @@ testRule(rule, {
     code: "a { padding-left: 10px; margin-right: 10px; margin-top: 20px; margin-bottom: 30px; }",
   }, {
     code: "a { padding-top: 1px; padding-bottom: 1px; } b { padding-left: 1px; padding-right: 1px; }",
+  }, {
+    code: "a { transition-delay: 0.5s; -webkit-transition-duration: 2s; -webkit-transition-property: top; -webkit-transition-timing-function: ease; }",
+    description: "one of properties is unprefixed",
+  }, {
+    code: "a { -webkit-transition-delay: 0.5s; transition-duration: 2s; transition-property: top; transition-timing-function: ease; }",
+    description: "one of properties is prefixed",
   } ],
 
   reject: [ {
@@ -38,6 +44,12 @@ testRule(rule, {
   }, {
     code: "a { border-top-width: 7px; border-top-style: double; border-top-color: green; }",
     message: messages.expected("border-top"),
+  }, {
+    code: "a { -webkit-transition-delay: 0.5s; -webkit-transition-duration: 2s; -webkit-transition-property: top; -webkit-transition-timing-function: ease; }",
+    message: messages.expected("-webkit-transition"),
+  }, {
+    code: "a { -WEBKIT-transition-delay: 0.5s; -webKIT-transition-duration: 2s; -webkit-TRANSITION-property: top; -webkit-transition-timing-function: ease; }",
+    message: messages.expected("-webkit-transition"),
   }, {
     code: "a { border-top-width: 1px; border-bottom-width: 1px; border-left-width: 1px; border-right-width: 1px; }",
     message: messages.expected("border-width"),