--- a/lib/rules/declaration-block-no-redundant-longhand-properties/index.js
+++ b/lib/rules/declaration-block-no-redundant-longhand-properties/index.js
@@ -36,8 +36,13 @@ const rule = function (actual, options) {
       })
     })
 
-    root.walkRules(check)
-    root.walkAtRules(check)
+    root.walk((node) => {
+      if (node.type !== "rule" && node.type !== "atrule") {
+        return
+      }
+
+      check(node)
+    })
 
     function check(statement) {
       const longhandDeclarations = {}