--- a/lib/rules/declaration-block-no-redundant-longhand-properties/index.js
+++ b/lib/rules/declaration-block-no-redundant-longhand-properties/index.js
@@ -2,6 +2,7 @@
 
 const _ = require("lodash")
 const optionsMatches = require("../../utils/optionsMatches")
+const postcss = require("postcss")
 const report = require("../../utils/report")
 const ruleMessages = require("../../utils/ruleMessages")
 const shorthandData = require("../../reference/shorthandData")
@@ -54,17 +55,29 @@ const rule = function (actual, options) {
         }
 
         const prop = node.prop.toLowerCase()
+        const unprefixedProp = postcss.vendor.unprefixed(prop)
+        const prefix = postcss.vendor.prefix(prop)
 
-        const shorthandProperties = longhandProperties[prop]
+        const shorthandProperties = longhandProperties[unprefixedProp]
 
         if (!shorthandProperties) {
           return
         }
 
         shorthandProperties.forEach(shorthandProperty => {
-          (longhandDeclarations[shorthandProperty] || (longhandDeclarations[shorthandProperty] = [])).push(prop)
+          const prefixedShorthandProperty = prefix + shorthandProperty
 
-          if (!_.isEqual(shorthandData[shorthandProperty].sort(), longhandDeclarations[shorthandProperty].sort())) {
+          if (!longhandDeclarations[prefixedShorthandProperty]) {
+            longhandDeclarations[prefixedShorthandProperty] = []
+          }
+
+          longhandDeclarations[prefixedShorthandProperty].push(prop)
+
+          const prefixedShorthandData = shorthandData[shorthandProperty].map((item) => {
+            return prefix + item
+          })
+
+          if (!_.isEqual(prefixedShorthandData.sort(), longhandDeclarations[prefixedShorthandProperty].sort())) {
             return
           }
 
@@ -72,7 +85,7 @@ const rule = function (actual, options) {
             ruleName,
             result,
             node,
-            message: messages.expected(shorthandProperty),
+            message: messages.expected(prefixedShorthandProperty),
           })
         })
       })