--- a/lib/rules/declaration-block-no-redundant-longhand-properties/index.js
+++ b/lib/rules/declaration-block-no-redundant-longhand-properties/index.js
@@ -37,13 +37,8 @@ const rule = function (actual, options) {
       })
     })
 
-    root.walk((node) => {
-      if (node.type !== "rule" && node.type !== "atrule") {
-        return
-      }
-
-      check(node)
-    })
+    root.walkRules(check)
+    root.walkAtRules(check)
 
     function check(statement) {
       const longhandDeclarations = {}