--- a/lib/rules/declaration-block-no-shorthand-property-overrides/README.md
+++ b/lib/rules/declaration-block-no-shorthand-property-overrides/README.md
@@ -48,7 +48,3 @@ a { padding: 10px; padding-left: 20px; }
 ```css
 a { transition-property: opacity; } a { transition: opacity 1s linear; }
 ```
-
-```css
-a { transition-property: opacity; -webkit-transition: opacity 1s linear; }
-```