--- a/lib/rules/declaration-block-no-shorthand-property-overrides/__tests__/index.js
+++ b/lib/rules/declaration-block-no-shorthand-property-overrides/__tests__/index.js
@@ -26,6 +26,8 @@ testRule(rule, {
   }, {
     code: "a { transition-property: opacity; } a { transition: opacity 1s linear; }",
   }, {
+    code: "a { -webkit-transition-property: opacity; transition: opacity 1s linear; }",
+  }, {
     code: "a { transition-property: opacity; -webkit-transition: opacity 1s linear; }",
   } ],
 
@@ -57,5 +59,8 @@ testRule(rule, {
   }, {
     code: "a { @media (color) { background-repeat: no-repeat; background: url(lion.png); }}",
     message: messages.rejected("background", "background-repeat"),
+  }, {
+    code: "a { -webkit-transition-property: opacity; -webkit-transition: opacity 1s linear; }",
+    message: messages.rejected("-webkit-transition", "-webkit-transition-property"),
   } ],
 })