--- a/lib/rules/declaration-block-no-shorthand-property-overrides/__tests__/index.js
+++ b/lib/rules/declaration-block-no-shorthand-property-overrides/__tests__/index.js
@@ -62,5 +62,8 @@ testRule(rule, {
   }, {
     code: "a { -webkit-transition-property: opacity; -webkit-transition: opacity 1s linear; }",
     message: messages.rejected("-webkit-transition", "-webkit-transition-property"),
+  }, {
+    code: "a { -WEBKIT-transition-property: opacity; -webKIT-transition: opacity 1s linear; }",
+    message: messages.rejected("-webKIT-transition", "-WEBKIT-transition-property"),
   } ],
 })