--- a/lib/rules/declaration-block-no-shorthand-property-overrides/index.js
+++ b/lib/rules/declaration-block-no-shorthand-property-overrides/index.js
@@ -1,5 +1,6 @@
 "use strict"
 
+const postcss = require("postcss")
 const report = require("../../utils/report")
 const ruleMessages = require("../../utils/ruleMessages")
 const shorthandData = require("../../reference/shorthandData")
@@ -34,22 +35,25 @@ const rule = function (actual) {
         if (node.type !== "decl") {
           return
         }
-        const prop = node.prop
 
-        const overrideables = shorthandData[prop.toLowerCase()]
+        const prop = node.prop
+        const unprefixedProp = postcss.vendor.unprefixed(prop)
+        const prefix = postcss.vendor.prefix(prop)
+
+        const overrideables = shorthandData[unprefixedProp.toLowerCase()]
         if (!overrideables) {
           declarations[prop.toLowerCase()] = prop
           return
         }
         overrideables.forEach(longhandProp => {
-          if (!declarations.hasOwnProperty(longhandProp)) {
+          if (!declarations.hasOwnProperty(prefix + longhandProp)) {
             return
           }
           report({
             ruleName,
             result,
             node,
-            message: messages.rejected(prop, declarations[longhandProp]),
+            message: messages.rejected(prop, declarations[prefix + longhandProp]),
           })
         })
       })