--- a/lib/rules/declaration-block-no-shorthand-property-overrides/index.js
+++ b/lib/rules/declaration-block-no-shorthand-property-overrides/index.js
@@ -38,7 +38,7 @@ const rule = function (actual) {
 
         const prop = node.prop
         const unprefixedProp = postcss.vendor.unprefixed(prop)
-        const prefix = postcss.vendor.prefix(prop)
+        const prefix = postcss.vendor.prefix(prop).toLowerCase()
 
         const overrideables = shorthandData[unprefixedProp.toLowerCase()]
         if (!overrideables) {