--- a/lib/rules/declaration-block-no-shorthand-property-overrides/index.js
+++ b/lib/rules/declaration-block-no-shorthand-property-overrides/index.js
@@ -19,13 +19,8 @@ const rule = function (actual) {
       return
     }
 
-    root.walk((node) => {
-      if (node.type !== "rule" && node.type !== "atrule") {
-        return
-      }
-
-      check(node)
-    })
+    root.walkRules(check)
+    root.walkAtRules(check)
 
     function check(statement) {
       const declarations = {}