--- a/lib/rules/shorthand-property-no-redundant-values/README.md
+++ b/lib/rules/shorthand-property-no-redundant-values/README.md
@@ -16,6 +16,7 @@ This rule warns you when you use redundant values in the following shorthand pro
 -   `border-radius`
 -   `border-style`
 -   `border-width`
+-   `grid-gap`
 
 ## Options