--- a/lib/rules/shorthand-property-no-redundant-values/index.js
+++ b/lib/rules/shorthand-property-no-redundant-values/index.js
@@ -5,7 +5,6 @@ const isStandardSyntaxProperty = require("../../utils/isStandardSyntaxProperty")
 const report = require("../../utils/report")
 const ruleMessages = require("../../utils/ruleMessages")
 const validateOptions = require("../../utils/validateOptions")
-const shorthandData = require("../../reference/shorthandData")
 const valueParser = require("postcss-value-parser")
 const postcss = require("postcss")
 
@@ -15,7 +14,14 @@ const messages = ruleMessages(ruleName, {
   rejected: (unexpected, expected) => `Unexpected longhand value '${unexpected}' instead of '${expected}'`,
 })
 
-const shorthandableProperties = new Set(Object.keys(shorthandData))
+const shorthandableProperties = new Set([
+  "margin",
+  "padding",
+  "border-color",
+  "border-radius",
+  "border-style",
+  "border-width",
+])
 
 const ignoredCharacters = [
   "+",
@@ -29,18 +35,6 @@ const ignoredCharacters = [
   "--",
   "var(",
 ]
-
-const ignoredShorthandProperties = new Set([
-  "background",
-  "font",
-  "border",
-  "border-top",
-  "border-bottom",
-  "border-left",
-  "border-right",
-  "list-style",
-  "transition",
-])
 
 function isIgnoredCharacters(value) {
   return ignoredCharacters.some(char => value.indexOf(char) !== -1)
@@ -112,7 +106,7 @@ const rule = function (actual) {
       const normalizedProp = postcss.vendor.unprefixed(prop.toLowerCase())
 
       // Ignore not shorthandable properties, and math operations
-      if (isIgnoredCharacters(value) || !shorthandableProperties.has(normalizedProp) || ignoredShorthandProperties.has(normalizedProp)) {
+      if (isIgnoredCharacters(value) || !shorthandableProperties.has(normalizedProp)) {
         return
       }