--- a/lib/rules/shorthand-property-no-redundant-values/index.js
+++ b/lib/rules/shorthand-property-no-redundant-values/index.js
@@ -21,6 +21,7 @@ const shorthandableProperties = new Set([
   "border-radius",
   "border-style",
   "border-width",
+  "grid-gap",
 ])
 
 const ignoredCharacters = [