--- a/lib/rules/shorthand-property-no-redundant-values/index.js
+++ b/lib/rules/shorthand-property-no-redundant-values/index.js
@@ -14,6 +14,7 @@ const messages = ruleMessages(ruleName, {
   rejected: (unexpected, expected) => `Unexpected longhand value '${unexpected}' instead of '${expected}'`,
 })
 
+// Only these shorthand properties can have values, that can be written in shorter form (remove repetitions)
 const shorthandableProperties = new Set([
   "margin",
   "padding",