<!---
Please read the following. Pull requests that do not adhere to these guidelines will be closed.

Each pull request must, with the exception of minor documentation fixes, be associated with an open issue. If a corresponding issue does not exist please stop. Instead, create an issue so we can discuss the change first.

If there is an associated open issue, then the next step is to make sure you've read the relevant developer guide:

- Adding a rule: https://github.com/stylelint/stylelint/blob/master/docs/developer-guide/rules.md#write-the-rule

- Adding an option: https://github.com/stylelint/stylelint/blob/master/docs/developer-guide/rules.md#adding-options-to-existing-rules

- Fixing a bug: https://github.com/stylelint/stylelint/blob/master/docs/developer-guide/rules.md#fixing-bugs

Once you've done that, then please continue by answering these two questions:  -->

> Which issue, if any, is this issue related to?

#1780 

> Is there anything in the PR that needs further explanation?

Added new shorthand properties from #1780, except `mask`, because [it's not clear](https://developer.mozilla.org/en-US/docs/Web/CSS/mask) what for this shorthand.

Removed prefixed shorthands. I don't think it's a good idea to keep list of all prefixed properties and shorthand.

I have concerns about `shorthand-property-no-redundant-values`. It uses `shorthandData` also. My concers is that it has a [list of ignored shorthands](https://github.com/stylelint/stylelint/blob/4ad8fbdbc4d126c618bd70d98bf94e4b207aefb8/lib/rules/shorthand-property-no-redundant-values/index.js#L33-L43). With new shorthands ignore list will be much bigger than list of shorthands supported in this rule. Maybe we should specify the list of allowed shorthands, rather using `shorthandData` and ignore.