@davidtheclark 2017-02-12
> except mask, because it's not clear what for this shorthand.