@davidtheclark 2017-02-12
> except mask, because it's not clear what for this shorthand.

Why is `mask` different from the others? I have never used it, but it's real! e.g. https://css-tricks.com/clipping-masking-css/

> Removed prefixed shorthands. I don't think it's a good idea to keep list of all prefixed properties and shorthand.

I agree, but I don't like that we're removing functionality for that. Can we use Autoprefixer in the rule to remove the need for a manual list?

> Maybe we should specify the list of allowed shorthands, rather using shorthandData and ignore.

Sounds good to me!

Want to add the changes above to this PR?