@hudochenkov 2017-02-12
> Why is mask different from the others? I have never used it, but it's real! e.g. https://css-tricks.com/clipping-masking-css/

I know it's real, but I don't know every property for this shorthand. [MDN](https://developer.mozilla.org/en-US/docs/Web/CSS/mask) and [W3C](https://drafts.fxtf.org/css-masking-1/#the-mask) don't make it clear. They say about `mask-clip`, `mask-origin`, and `mask-border` only, but `mask` accept more than this properties, as I see.

> I agree, but I don't like that we're removing functionality for that. Can we use Autoprefixer in the rule to remove the need for a manual list?

I will look how `declaration-block-no-redundant-longhand-properties`, `declaration-block-no-shorthand-property-overrides`, and `shorthand-property-no-redundant-values` work and see what can be done.

> > Maybe we should specify the list of allowed shorthands, rather using shorthandData and ignore.

> Sounds good to me!

> Want to add the changes above to this PR?

Sure!