@davidtheclark 2017-02-12
@hudochenkov This page looks helpful: https://tympanus.net/codrops/css_reference/mask/