@hudochenkov 2017-02-16
Benchmarks for https://github.com/stylelint/stylelint/pull/2354/commits/ac40d7a4a2b82bbd72837b7911cccfa31e4189a3

Before:

```
> stylelint@7.8.0 benchmark-rule /Users/aleks/projects/stylelint
> node scripts/benchmark-rule.js "declaration-block-no-shorthand-property-overrides" "true"

Warnings: 0
Mean: 45.400157870370364 ms
Deviation: 11.06851959653398 ms

Mean: 41.5444920952381 ms
Deviation: 11.320279370467999 ms

Mean: 41.935588403508774 ms
Deviation: 10.520627666926606 ms

Mean: 41.81118664912281 ms
Deviation: 11.763514223194033 ms

Mean: 41.28931149206349 ms
Deviation: 10.750820788142141 ms
```

After:

```
Warnings: 0
Mean: 40.35173133620689 ms
Deviation: 10.237891375944221 ms

Mean: 43.27671090243902 ms
Deviation: 11.477326149577069 ms

Mean: 40.06621774358975 ms
Deviation: 11.043813336306489 ms

Mean: 40.419671480620174 ms
Deviation: 12.434897384556843 ms

Mean: 42.71238819642859 ms
Deviation: 9.455533365058331 ms

Mean: 39.48838472435897 ms
Deviation: 10.162578651704262 ms
```